Remove explicit compile entries
